Major code quality improvements and bug fixes
Critical fixes:
- Fixed duplicate parameter IDs for address 100 (removed invalid 3138/3139)
- Added pthread mutex protection for mqtt_connected flag (thread safety)
- Fixed response validation to properly re-send requests on mismatch
- Added NULL check for mosquitto_new()

Production readiness:
- Added signal handlers (SIGINT/SIGTERM) for graceful shutdown
- Fixed mosquitto_loop_stop() blocking issue (force=true)
- Defined all magic numbers as constants
- Improved VMIN/VTIME timeout documentation

Code quality:
- Removed unused on_publish callback
- Added NUM_PARAMETERS macro for safer array iteration
- Fixed all compiler warnings (added -Wall -Wextra)
- Improved type safety with explicit casts
- Consistent error handling with serial_flush()

Build system:
- Optimized Makefile to share library objects between builds
- Only main.c compiles twice (normal/debug), ~50% faster builds
- Added -O2 optimization for production binary
- Added -pthread and -lpthread flags
- Zero compiler warnings

+++ src/serial.c
@@ -79,9 +79,12 @@ static int set_interface_attribs(int fd, int speed, serial_parity_t parity, int 
     tio.c_lflag = 0; // will be noncanonical mode (ICANON not set)
 
     // set blocking with timeout
-    // VMIN = 0 and VTIME > 0: read returns as soon as data is available or timeout occurs
+    // VMIN = 0 and VTIME > 0: read() returns immediately when data arrives,
+    // or after VTIME deciseconds if no data received.
+    // NOTE: Timer resets on each received byte, so total frame time can exceed VTIME.
+    // The serial_read() function loops to collect full frames.
     tio.c_cc[VMIN] = 0;   // minimum number of characters for noncanonical read
-    tio.c_cc[VTIME] = 20; // 2 second read timeout (in deciseconds)
+    tio.c_cc[VTIME] = 20; // 2 second timeout per read() call (in deciseconds)
 
     if (tcsetattr(fd, TCSANOW, &tio) != 0) {
         error_message("tcsetattr error %d: %s\n", errno, strerror(errno));
@@ -124,7 +127,7 @@ int serial_write(const void *ptr, unsigned size) {
     
     if (bytes_written < 0) {
         error_message("Write error %d: %s\n", errno, strerror(errno));
-    } else if (bytes_written != size) {
+    } else if ((unsigned int)bytes_written != size) {
         SERIAL_DEBUG_PRINT("Warning: Only wrote %d of %u bytes\n", bytes_written, size);
     } else {
         SERIAL_DEBUG_PRINT("Successfully wrote %d bytes\n", bytes_written);
